+++ src/backend/defs/go-objects/user.go
@@ -44,6 +44,12 @@ type PasswordResetRequest struct {
 	Username     string `json:"username"`
 }
 
+// payload for triggering delete user process inside an `actions` call
+type DeleteUserParams struct {
+	Password    string `json:"password"`
+	AccessToken string `json:"access_token"`
+}
+
 // data stored into cache for authenticated user
 type Auth_cache struct {
 	Access_token  string    `json:"access_token"`
@@ -55,14 +61,6 @@ type Auth_cache struct {
 	Y             big.Int   `json:"y"`
 	Key_id        string    `json:"key_id"`
 	Shard_id      string    `json:"shard_id"`
-}
-
-// data stored into cache as long as a reset password request is pending for an user
-type Pass_reset_session struct {
-	Reset_token string    `json:"reset_token"`
-	Expires_in  int       `json:"expires_in"`
-	Expires_at  time.Time `json:"expires_at"`
-	User_id     string    `json:"user_id"`
 }
 
 // unmarshal a map[string]interface{} that must owns all Contact's fields