+++ src/frontend/web_application/src/store/modules/device.js
@@ -11,6 +11,7 @@ export const CREATE_DEVICE = 'co/device/CREATE_DEVICE';
 export const UPDATE_DEVICE = 'co/device/UPDATE_DEVICE';
 export const REMOVE_DEVICE = 'co/device/REMOVE_DEVICE';
 export const VERIFY_DEVICE = 'co/device/VERIFY_DEVICE';
+export const VALIDATE_DEVICE = 'co/device/VALIDATE_DEVICE';
 
 export function setNewDevice(isNew) {
   return {
@@ -89,7 +90,26 @@ export function verifyDevice({ device }) {
   return {
     type: VERIFY_DEVICE,
     payload: {
-      device,
+      request: {
+        method: 'post',
+        url: `/api/v2/devices/${device.device_id}/actions`,
+        data: {
+          actions: ['device-validation'],
+          params: { channel: 'email' },
+        },
+      },
+    },
+  };
+}
+
+export function requestDeviceVerification({ token }) {
+  return {
+    type: VALIDATE_DEVICE,
+    payload: {
+      request: {
+        method: 'get',
+        url: `/api/v2/validate-device/${token}`,
+      },
     },
   };
 }